move the cross account lambda to deliver_to_firehose lambda
